Struct wasmer_engine_universal::engine::UniversalEngineInner

pub struct UniversalEngineInner {
    compiler: Option<Box<dyn Compiler>>,
    features: Features,
    code_memory: Vec<CodeMemory>,
    signatures: SignatureRegistry,
    func_data: Arc<FuncDataRegistry>,
}

The inner contents of UniversalEngine

Fields
compiler: Option<Box<dyn Compiler>>

The compiler

features: Features

The features to compile the Wasm module with

code_memory: Vec<CodeMemory>

The code memory is responsible of publishing the compiled functions to memory.

signatures: SignatureRegistry

The signature registry is used mainly to operate with trampolines performantly.

func_data: Arc<FuncDataRegistry>

The backing storage of VMFuncRefs. This centralized store ensures that 2 functions with the same VMCallerCheckedAnyfunc will have the same VMFuncRef. It also guarantees that the VMFuncRefs stay valid until the engine is dropped.

Implementations

impl UniversalEngineInner

pub fn compiler(&self) -> Result<&dyn Compiler, CompileError>

Gets the compiler associated to this engine.

pub fn validate<'data>(&self, data: &'data [u8]) -> Result<(), CompileError>

Validate the module

pub fn features(&self) -> &Features

The Wasm features

pub(crate) fn allocate(
&mut self,
_module: &ModuleInfo,
functions: &PrimaryMap<LocalFunctionIndex, FunctionBody>,
function_call_trampolines: &PrimaryMap<SignatureIndex, FunctionBody>,
dynamic_function_trampolines: &PrimaryMap<FunctionIndex, FunctionBody>,
custom_sections: &PrimaryMap<SectionIndex, CustomSection>
) -> Result<(PrimaryMap<LocalFunctionIndex, FunctionExtent>, PrimaryMap<SignatureIndex, VMTrampoline>, PrimaryMap<FunctionIndex, FunctionBodyPtr>, PrimaryMap<SectionIndex, SectionBodyPtr>), CompileError>

Allocate compiled functions into memory

pub(crate) fn publish_compiled_code(&mut self)

Make memory containing compiled code executable.

pub(crate) fn publish_eh_frame(
&mut self,
eh_frame: Option<&[u8]>
) -> Result<(), CompileError>

Register DWARF-type exception handling information associated with the code.

pub fn signatures(&self) -> &SignatureRegistry

Shared signature registry.

pub(crate) fn func_data(&self) -> &Arc<FuncDataRegistry>

Shared func metadata registry.