Struct wasmer_engine_universal::serialize::SerializableModule

pub struct SerializableModule {
    pub compilation: SerializableCompilation,
    pub compile_info: CompileModuleInfo,
    pub data_initializers: Box<[OwnedDataInitializer]>,
}

Serializable struct that is able to serialize from and to a UniversalArtifactInfo.

Fields
compilation: SerializableCompilationcompile_info: CompileModuleInfodata_initializers: Box<[OwnedDataInitializer]>
Implementations

impl SerializableModule

pub fn serialize(&self) -> Result<Vec<u8>, SerializeError>

Serialize a Module into bytes The bytes will have the following format: RKYV serialization (any length) + POS (8 bytes)

pub unsafe fn deserialize(
metadata_slice: &[u8]
) -> Result<Self, DeserializeError>

Deserialize a Module from a slice. The slice must have the following format: RKYV serialization (any length) + POS (8 bytes)

Safety

This method is unsafe since it deserializes data directly from memory. Right now we are not doing any extra work for validation, but rkyv has an option to do bytecheck on the serialized data before serializing (via rkyv::check_archived_value).

unsafe fn archive_from_slice<'a>(
metadata_slice: &'a [u8]
) -> Result<&'a ArchivedSerializableModule, DeserializeError>

Safety

This method is unsafe. Please check SerializableModule::deserialize for more details.

pub fn deserialize_from_archive(
archived: &ArchivedSerializableModule
) -> Result<Self, DeserializeError>