Struct wasmer_engine_universal::UniversalArtifact

pub struct UniversalArtifact { /* private fields */ }

A compiled wasm module, ready to be instantiated.

Implementations

impl UniversalArtifact

pub fn new(
engine: &UniversalEngine,
data: &[u8],
tunables: &dyn Tunables
) -> Result<Self, CompileError>

Compile a data buffer into a UniversalArtifactBuild, which may then be instantiated.

pub unsafe fn deserialize(
engine: &UniversalEngine,
bytes: &[u8]
) -> Result<Self, DeserializeError>

Deserialize a UniversalArtifactBuild

Safety

This function is unsafe because rkyv reads directly without validating the data.

pub fn from_parts(
engine_inner: &mut UniversalEngineInner,
artifact: UniversalArtifactBuild
) -> Result<Self, CompileError>

Construct a UniversalArtifactBuild from component parts.

pub fn get_default_extension(triple: &Triple) -> &'static str

Get the default extension when serializing this artifact

pub fn is_deserializable(bytes: &[u8]) -> bool

Check if the provided bytes look like a serialized UniversalArtifactBuild.