Crate wasmer_engine_universal

Universal backend for Wasmer compilers.

Given a compiler (such as CraneliftCompiler or LLVMCompiler) it generates the compiled machine code, and publishes it into memory so it can be used externally.

Modules

artifact
Define UniversalArtifact to allow compiling and instantiating to be done as separate steps.

builder

code_memory
Memory management for executable code.

engine
Universal compilation.

link
Linking for Universal-compiled code.

serialize

unwind

Structs

CodeMemory
Memory manager for executable code.

Universal
The Universal builder

UniversalArtifact
A compiled wasm module, ready to be instantiated.

UniversalEngine
A WebAssembly Universal Engine.

Constants

VERSION
Version number of this crate.

Functions

link_module
Links a module, patching the allocated functions with the required relocations and jump tables.